fix python issues

+++ blueprints/secops-anonymization-pipeline/source/shared/utils.py
@@ -13,16 +13,19 @@
 # limitations under the License.
 #
 
-import os
+import csv
 import logging
 import math
-import csv
+import os
+import re
+from datetime import datetime, time, timedelta, timezone
+
 from google.cloud import storage
-from datetime import datetime, timedelta, timezone, time
 
 LOGGER = logging.getLogger("secops")
 """Utility functions required for ingestion scripts."""
 MAX_FILE_SIZE = 61440000  # Max size supported by DLP
+ANONYMIZED_FOLDER_REGEX = r"^([^/]+)/([^/]+)/"
 
 
 def format_date_time_range(date_input):
@@ -38,7 +41,7 @@ def format_date_time_range(date_input):
             - Start of day: "YYYY-MM-DDTHH:MM:SSZ"
             - End of day:   "YYYY-MM-DDTHH:MM:SSZ"
     """
-    date_obj = datetime.strptime(date_input, "%Y-%m-%d")
+    date_obj = datetime.strptime(date_input, "%Y-%m-%d").replace(tzinfo=timezone.utc)
 
     start_of_day = datetime.combine(date_obj.date(), time.min, tzinfo=timezone.utc)
     end_of_day = start_of_day + timedelta(days=1, seconds=-1)
@@ -47,23 +50,27 @@ def format_date_time_range(date_input):
 
 
 def list_anonymized_folders(bucket_name, folder_name):
-    """Lists all folders (prefixes) within a specified folder in a GCS bucket.
+    """Lists folders matching a specific date pattern in a Google Cloud Storage bucket.
 
     Args:
-        bucket_name: Name of the GCS bucket.
-        folder_name: Name of the folder (prefix) to search within.
+      bucket_name: The name of the GCS bucket.
+      folder_name: Target folder inside the GCS bucket.
 
     Returns:
-        A list of folder names (prefixes) found.
+      A list of folder names (strings) matching the date pattern.
     """
-    folders = []
     storage_client = storage.Client()
-    for blob in storage_client.list_blobs(bucket_name, prefix=f"{folder_name}/"):
-        folder_name = blob.name.split("/")[1]
-        if folder_name not in folders:
-            folders.append(folder_name)
+    bucket = storage_client.bucket(bucket_name)
 
-    return folders
+    blobs = bucket.list_blobs(prefix=f"{folder_name}/")
+
+    folders = set()
+    for blob in blobs:
+        match = re.match(ANONYMIZED_FOLDER_REGEX, blob.name)
+        if match:
+            folders.add(match.group(2))
+
+    return list(folders)
 
 
 def delete_folder(bucket_name, folder_name):
@@ -90,7 +97,6 @@ def list_log_files(bucket_name, folder_name):
     Returns:
         A list of folder names (prefixes) found.
     """
-
     storage_client = storage.Client()
     csv_files = []
     for blob in storage_client.list_blobs(bucket_name, prefix=f"{folder_name}/"):
@@ -100,13 +106,44 @@ def list_log_files(bucket_name, folder_name):
     return csv_files
 
 
-def split_csv(bucket_name, blob_name, file_size):
-    """Splits a CSV file into smaller chunks and uploads them back to the bucket.
+def upload_single_csv_as_log(bucket_name, blob_name):
+    storage_client = storage.Client()
+    bucket = storage_client.bucket(bucket_name)
+    blob = bucket.blob(blob_name)
+    new_name = blob_name.replace(".csv", ".log")
+    bucket.rename_blob(blob, new_name)
+
+
+def split_and_rename_csv_to_log_files(bucket_name, folder_name):
+    """Splits all CSV files in a GCS folder into chunks if they exceed the maximum size,
+    converts each chunk into a single-column CSV, and uploads the chunks as .log
+    files.
 
     Args:
       bucket_name: The name of the GCS bucket.
-      blob_name: The name of the CSV blob in the bucket.
-      max_file_size: The maximum size of each chunk in bytes.
+      folder_name: The name of the folder containing the CSV files.
+    """
+    storage_client = storage.Client()
+    bucket = storage_client.bucket(bucket_name)
+
+    blobs = bucket.list_blobs(prefix=f"{folder_name}/")
+
+    for blob in blobs:
+        if blob.name.endswith(".csv"):
+            file_size = blob.size
+            if file_size > MAX_FILE_SIZE:
+                split_and_upload_csv(bucket_name, blob.name, file_size)
+            else:
+                upload_single_csv_as_log(bucket_name, blob.name)
+
+
+def split_and_upload_csv(bucket_name, blob_name, file_size):
+    """Splits a large CSV file from GCS into smaller chunks and uploads them as .log files.
+
+    Args:
+      bucket_name: The name of the GCS bucket.
+      blob_name: The name of the CSV blob in GCS.
+      file_size: The size of the CSV file in bytes.
     """
     storage_client = storage.Client()
     bucket = storage_client.bucket(bucket_name)
@@ -116,8 +153,8 @@ def split_csv(bucket_name, blob_name, file_size):
     temp_file = "/tmp/temp.csv"
     blob.download_to_filename(temp_file)
 
-    file = open(temp_file, encoding="utf8")
-    numline = sum(1 for row in csv.reader(file))
+    with open(temp_file, encoding="utf8") as file:
+        numline = sum(1 for _ in csv.reader(file))
 
     # Read the CSV file in chunks
     chunk_number = math.ceil(numline * MAX_FILE_SIZE / file_size)
@@ -139,16 +176,15 @@ def split_csv(bucket_name, blob_name, file_size):
                 index += 1
                 lines = []
 
-        chunk_filename = f"{blob_name.split('.')[0]}_{index}.log"
-        chunk_path = f"/tmp/temp-{index}.csv"
-        with open(chunk_path, "w") as fout:
-            fout.writelines(lines)
-        chunk_blob = bucket.blob(f"{chunk_filename}")
-        chunk_blob.upload_from_filename(chunk_path)
-        print(f"Uploaded {chunk_filename} to {bucket_name}")
-        os.remove(chunk_path)  # Remove the local chunk file
-        index += 1
-        lines = []
+        if lines:
+            chunk_filename = f"{blob_name.split('.')[0]}_{index}.log"
+            chunk_path = f"/tmp/temp-{index}.csv"
+            with open(chunk_path, "w") as fout:
+                fout.writelines(lines)
+            chunk_blob = bucket.blob(f"{chunk_filename}")
+            chunk_blob.upload_from_filename(chunk_path)
+            print(f"Uploaded {chunk_filename} to {bucket_name}")
+            os.remove(chunk_path)  # Remove the local chunk file
 
     # Remove the temporary file
     os.remove(temp_file)
@@ -156,26 +192,6 @@ def split_csv(bucket_name, blob_name, file_size):
     # remove old log file
     blob = bucket.blob(blob_name)
     blob.delete()
-
-
-def split_and_rename_csv_to_log_files(bucket_name, folder_name):
-    """Renames all .csv files to .log files within a GCS bucket folder (and subfolders).
-
-    Args:
-        bucket_name (str): Name of the GCS bucket.
-        folder_prefix (str): Prefix of the folder within the bucket to process.
-    """
-
-    storage_client = storage.Client()
-    bucket = storage_client.bucket(bucket_name)
-
-    blobs = storage_client.list_blobs(bucket, prefix=f"{folder_name}/")
-    for blob in blobs:
-        if blob.name.endswith(".csv") and blob.size >= MAX_FILE_SIZE:
-            split_csv(bucket_name, blob.name, blob.size)
-        elif blob.name.endswith(".csv"):
-            new_name = blob.name.replace(".csv", ".log")
-            bucket.rename_blob(blob, new_name)
 
 
 def get_secops_export_folders_for_date(bucket_name, export_date):
